refactor: move out Outputdialog to own file
